+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## Unreleased
+
 ## [2.1.0] - 2019-06-25
 - Added getPathAndQueryString method to Url class.