+++ CHANGELOG.md
@@ -1,30 +1,42 @@
 ## Unreleased
+
 - Added withFilename method to FilePath and UrlPath classes.
 - Added parseAsDirectory and tryParseAsDirectory methods to FilePath and UrlPath classes.
+- Changed required PHP version to >= 7.2.
 
 ## [2.1.0] - 2019-06-25
+
 - Added getPathAndQueryString method to Url class.
 
 ## [2.0.0] - 2017-12-25
+
 - Added type hints to all data type classes.
 - Changed required PHP version to >= 7.1.
 
 ## [1.2.0] - 2017-12-22
+
 - Added equals method to all data type classes.
 - Added toInteger method to IPAddress class.
 - Added fromInteger method to IPAddress class.
 
 ## [1.1.0] - 2017-10-01
+
 - Added EmailAddress class.
 
 ## [1.0.1] - 2017-05-29
+
 - Fixed [#4](https://github.com/themichaelhall/datatypes/issues/4) - Url with name and/or password should not be treated as invalid.
 
 ## 1.0.0 - 2017-05-23
+
 - First stable revision.
 
 [2.1.0]: https://github.com/themichaelhall/datatypes/compare/v2.0.0...v2.1.0
+
 [2.0.0]: https://github.com/themichaelhall/datatypes/compare/v1.2.0...v2.0.0
+
 [1.2.0]: https://github.com/themichaelhall/datatypes/compare/v1.1.0...v1.2.0
+
 [1.1.0]: https://github.com/themichaelhall/datatypes/compare/v1.0.1...v1.1.0
+
 [1.0.1]: https://github.com/themichaelhall/datatypes/compare/v1.0.0...v1.0.1