+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## [2.3.0] - 2021-02-10
 
 - Changed required PHP version to >= 7.3.
 
@@ -35,6 +35,7 @@
 
 - First stable revision.
 
+[2.3.0]: https://github.com/themichaelhall/datatypes/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/themichaelhall/datatypes/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/themichaelhall/datatypes/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/themichaelhall/datatypes/compare/v1.2.0...v2.0.0